+++ seniordesignlab1/.firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "arduinotempsensor-22a0f"
+  }
+}
