+++ seniordesignlab1/database.rules.json
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    ".read": true,  // 2021-10-28
+    ".write": true,  // 2021-10-28
+  }
+}
