--- seniordesignlab1/index.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Getting Started with Chart JS with www.chartjs3.com</title>
-    <link rel="stylesheet" href="main.css">
-  </head>
-  <body>
-    <div class="chartCard">
-      <div id="chartBox1" class="chartBox1">
-        <button onclick="toFarenheight()">Farenheight</button>
-        <button onclick="turnOnDisplay()">Turn on Display</button>
-        <canvas id="myChart1"></canvas>        
-      </div>
-
-      <div id="chartBox2" class="chartBox2">
-        <button onclick="toCelsius()">Celsius</button>
-        <canvas id="myChart2"></canvas>        
-      </div>
-      
-    </div>
-
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/hammer.js/2.0.8/hammer.min.js" 
-    integrity="sha512-UXumZrZNiOwnTcZSHLOfcTs0aos2MzBWHXOHOuB0J/R44QB0dwY5JgfbvljXcklVf65Gc4El6RjZ+lnwd2az2g==" 
-    crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/chartjs-plugin-zoom/1.1.1/chartjs-plugin-zoom.min.js" 
-    integrity="sha512-NxlWEbNbTV6acWnTsWRLIiwzOw0IwHQOYUCKBiu/NqZ+5jSy7gjMbpYI+/4KvaNuZ1qolbw+Vnd76pbIUYEG8g==" 
-    crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-
-    <script src="app.js"></script>
-
-  </body>
-</html>
